Code update: minor changes

+++ src/main/java/com/backend/liaison_system/users/admin/dao/StudentLocationData.java
@@ -3,7 +3,7 @@ package com.backend.liaison_system.users.admin.dao;
 public record StudentLocationData(
         StudentDetails studentDetails,
         CompanyDetails companyDetails,
-        int lat,
-        int lng
+        double lat,
+        double lng
 ) {
 }
